refactor: moving styling around
