+++ src/design-system/css/components/index.css
@@ -2,3 +2,4 @@
 @import "Table";
 @import "Input";
 @import "Dropdown";
+@import "FoodForm";
